backend/requirements.txt
fastapi
uvicorn
python-multipart
networkx
aiofiles
pytest
httpx
python-dotenv
